+++ src/test/java/com/challenge/verifier/placeOrder/domain/EventTest.java
@@ -4,7 +4,6 @@ import com.challenge.verifier.placeOrder.helper.TestOrderBuilder;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.List;
 
@@ -28,7 +27,7 @@ public class EventTest {
         EventPersistentModel eventPersistentModel = new Event(order, EventType.ORDER_PLACED, now).asPersistentModel();
         assertEquals(eventPersistentModel.getId(), 10000L);
         assertEquals(eventPersistentModel.getSide(), Side.BUY.name());
-        assertEquals(eventPersistentModel.getPrice(), BigDecimal.valueOf(98));
+        assertEquals(eventPersistentModel.getPrice(), 98);
         assertEquals(eventPersistentModel.getQuantity(), 25500);
         assertEquals(eventPersistentModel.getEvent_type(), EventType.ORDER_PLACED.name());
         assertEquals(eventPersistentModel.getEvent_time(), now);
